+++ Sources/SwiftlyPDFKit/Core/HTMLUtilities.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+// MARK: - HTML Escaping
+
+extension String {
+    /// Escapes characters that have special meaning in HTML.
+    var htmlEscaped: String {
+        self.replacingOccurrences(of: "&", with: "&amp;")
+            .replacingOccurrences(of: "<", with: "&lt;")
+            .replacingOccurrences(of: ">", with: "&gt;")
+            .replacingOccurrences(of: "\"", with: "&quot;")
+    }
+}